Update nvram.md
This isn't actually an update, but a request for some extra information in this section. I was unable to follow this page completely, and have been trying to find the answers to my questions elsewhere. It would be great if this section could be expanded to be more clear to those reading it. Thanks for considering!